+++ CHANGELOG.md
@@ -10,13 +10,14 @@ Scribe-Data tries to follow [semantic versioning](https://semver.org/), a MAJOR.
 
 Emojis for the following are chosen based on [gitmoji](https://gitmoji.dev/).
 
-# Scribe-Data 1.1.0
+# Scribe-Data 2.0.0
 
 ### ✨ Features
 
-<!-- - Scribe-Data now has the ability to download Wikipedia dumps of any language.
+- Scribe-Data now has the ability to download Wikipedia dumps of any language.
 - Processes have been added to parse and clean the above dumps.
-- Autosuggestions are generated from the cleaned texts by deriving most common words and those words that most commonly follow them. -->
+- Autosuggestions are generated from the cleaned texts by deriving most common words and those words that most commonly follow them.
+<!-- - The adjectives column has been removed from Scribe data tables until support is offered. -->
 
 ### ♻️ Code Refactoring